bigint for login

+++ Alarm_db.sql
@@ -7,7 +7,7 @@ USE alarm_db;
 CREATE TABLE users
 (
     id int NOT NULL AUTO_INCREMENT,
-    u_login int not null unique,
+    u_login bigint not null unique,
     u_passhash varchar(256) not null,
     u_name varchar(256),
     CONSTRAINT PK_users PRIMARY KEY(id)
